wip moving consensus to a single place
